Rewrited all client part, now it's modular. Also rewrited server, voice part broke. Now in process of making identification in server
