+++ app/blueprints/templates/wav_to_text.html
@@ -6,7 +6,7 @@
     <title>WAV to Text</title>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/4.4.1/socket.io.js"></script>
     <link rel="stylesheet" href="{{ url_for('static', filename='css/style.css') }}">
-    <script src="{{ url_for('static', filename='js/config.js') }}"></script>
+    <script src="{{ url_for('static', filename='js/configs/config.js') }}"></script>
     <script type="module" src="{{ url_for('static', filename='js/wav_to_text.js') }}" defer></script>
 </head>
 
@@ -35,7 +35,7 @@
         <div id="resultForm">
             <h2>Результат транскрибирования</h2>
             <div id="result"></div>
-            <button id="downloadBtn" style="display: none; align-self: end; margin-right: 30px;" onclick="downloadText()">Скачать</button>
+            <button id="downloadBtn" style="display: none; align-self: end; margin-right: 30px;">Скачать</button>
         </div>
     </div>
 </body>
