+++ app/config.py
@@ -1,0 +1,15 @@
+import json
+import os
+
+class Config:
+    SECRET_KEY = 'secret!'
+    RECORDS_DIR = 'records'
+    UPLOADS_DIR = 'uploads'
+
+    # Загрузка конфигурации из config.json
+    with open("app/static/config.json") as config_file:
+        config = json.load(config_file)
+
+    SERVER_URI = config['serverUrl']
+    VOSK_URI = config['voskUrl']
+    os.makedirs(RECORDS_DIR, exist_ok=True)
