+++ app/extensions.py
@@ -1,0 +1,3 @@
+from flask_socketio import SocketIO
+
+socketio = SocketIO(cors_allowed_origins="*", manage_session=True)
