+++ app/static/js/modules/dataConversionProcessor.js
