+++ app/static/js/modules/webSocketHandler.js
