+++ requirements.txt
@@ -1,0 +1,5 @@
+Flask
+Flask-SocketIO
+Flask-CORS
+eventlet
+websockets
